made the kernel class header file
